+++ js/util.js
@@ -1,0 +1,31 @@
+/*eslint-disable no-alert*/
+const getRandomIntInclusive = (min, max) => {
+  if (min < 0) {
+    alert('Поменяйте первый параметр, значение минимума');
+    return NaN;
+  }
+  if (max < 0) {
+    alert('Поменяйте второй параметр, значение максимума');
+    return NaN;
+  }
+  if (max <= min) {
+    alert('Поменяйте один из параметров, чтобы минимум был меньше максимума');
+    return NaN;
+  }
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  const result = Math.random() * (max - min + 1) + min;
+  return Math.floor(result);
+};
+
+/*
+https://developer.mozilla.org/ru/docs/Web/JavaScript/Reference/Global_Objects/Math/random
+
+function getRandomIntInclusive(min, max) {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+http://kodesource.top/javascript/form/string-length.php
+*/
+export {getRandomIntInclusive};
